Move TypeModifiers into own file and add some more tests
