+++ sql/src/common.rs
@@ -11,13 +11,14 @@ pub use literal::Literal;
 mod datatype;
 pub use datatype::DataType;
 
+mod typemodifiers;
+pub use typemodifiers::TypeModifier;
+
 mod value;
 pub use value::ValueExpression;
 
 mod aggregates;
 pub use aggregates::AggregateExpression;
-
-use crate::Parser as _Parser;
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum BinaryOperator {
@@ -176,60 +177,6 @@ where
     }
 }
 
-#[derive(Debug, PartialEq, Clone)]
-pub enum TypeModifier {
-    NotNull,
-    PrimaryKey,
-    Null,
-    DefaultValue { value: Option<Literal<'static>> },
-    Collate { collation: String },
-    Sequence { name: String },
-}
-
-impl<'i> crate::Parser<'i> for TypeModifier {
-    fn parse() -> impl Fn(&'i [u8]) -> IResult<&'i [u8], Self, nom::error::VerboseError<&'i [u8]>> {
-        move |i| {
-            #[allow(deprecated)]
-            type_modifier(i)
-        }
-    }
-}
-
-#[deprecated]
-pub fn type_modifier(i: &[u8]) -> IResult<&[u8], TypeModifier, nom::error::VerboseError<&[u8]>> {
-    nom::branch::alt((
-        nom::sequence::tuple((
-            nom::bytes::complete::tag_no_case("PRIMARY"),
-            nom::character::complete::multispace1,
-            nom::bytes::complete::tag_no_case("KEY"),
-        ))
-        .map(|_| TypeModifier::PrimaryKey),
-        nom::sequence::tuple((
-            nom::bytes::complete::tag_no_case("NOT"),
-            nom::character::complete::multispace1,
-            nom::bytes::complete::tag_no_case("NULL"),
-        ))
-        .map(|_| TypeModifier::NotNull),
-        nom::bytes::complete::tag_no_case("NULL").map(|_| TypeModifier::Null),
-        nom::sequence::tuple((
-            nom::bytes::complete::tag_no_case("DEFAULT"),
-            nom::character::complete::multispace1,
-            nom::combinator::opt(Literal::parse()),
-        ))
-        .map(|(_, _, value)| TypeModifier::DefaultValue {
-            value: value.map(|v| v.to_static()),
-        }),
-        nom::sequence::tuple((
-            nom::bytes::complete::tag_no_case("COLLATE"),
-            nom::character::complete::multispace1,
-            Identifier::parse(),
-        ))
-        .map(|(_, _, ident)| TypeModifier::Collate {
-            collation: ident.0.to_string(),
-        }),
-    ))(i)
-}
-
 #[cfg(test)]
 mod tests {
     use crate::macros::parser_parse;
@@ -266,17 +213,6 @@ mod tests {
             ColumnReference {
                 relation: None,
                 column: Identifier(Cow::Borrowed("column")),
-            }
-        );
-    }
-
-    #[test]
-    fn type_modifier_testing() {
-        parser_parse!(
-            TypeModifier,
-            "COLLATE \"C\"",
-            TypeModifier::Collate {
-                collation: "C".into()
             }
         );
     }
